back/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pyjwt[crypto]==2.8.0
httpx==0.27.0
pydantic==2.7.4
sqlmodel==0.0.22
SQLAlchemy==2.0.36

